refactor(website): emit safe rich text vnodes directly
